# Epic E068 Phase 1: Foundation Performance Optimization - 2× Performance Improvement