--- a/dbt/models/analysis/validate_enrollment_architecture.sql
+++ b/dbt/models/analysis/validate_enrollment_architecture.sql
@@ -1,4 +1,7 @@
-{{ config(materialized='table') }}
+{{ config(
+    materialized='table',
+    tags=['VALIDATION']
+) }}
 
 /*
   Enrollment Architecture Validation