--- a/dbt/models/data_quality/dq_integrity_summary.sql
+++ b/dbt/models/data_quality/dq_integrity_summary.sql
@@ -1,6 +1,6 @@
 {{ config(
     materialized='table',
-    tags=['data_quality', 'monitoring', 'summary']
+    tags=['VALIDATION', 'data_quality', 'monitoring', 'summary']
 ) }}
 
 -- Data Quality Summary: High-level view of integrity status