… .claude/settings.local.json
… .github/workflows/e068h-validation.yml
… .github/workflows/performance-benchmark.yml
… .navigator_checkpoints/year_2025.json
… .navigator_checkpoints/year_2026.json
… .navigator_checkpoints/year_2027.json
… .navigator_checkpoints/year_2028.json
… .navigator_checkpoints/year_2029.json
… E068F_IMPLEMENTATION_SUMMARY.md
… Jenkinsfile.e068h
… artifacts/runs/20250903_151613-b9d31eab/performance.json
… artifacts/runs/20250903_151613-b9d31eab/summary.json
… artifacts/runs/20250903_153310-d362e215/performance.json
… artifacts/runs/20250903_153310-d362e215/summary.json
… artifacts/runs/20250903_192336-e0f56be6/performance.json
… artifacts/runs/20250903_192336-e0f56be6/summary.json
… artifacts/runs/20250903_194202-e25e5140/performance.json
… artifacts/runs/20250903_194202-e25e5140/summary.json
… artifacts/runs/20250904_073639-0c254899/performance.json
… artifacts/runs/20250904_073639-0c254899/summary.json
… artifacts/runs/20250904_080736-15713676/performance.json
… artifacts/runs/20250904_080736-15713676/summary.json
… artifacts/runs/20250904_081701-432309fa/performance.json
… artifacts/runs/20250904_081701-432309fa/summary.json
… artifacts/runs/20250904_115056-276cf8b5/performance.json
… artifacts/runs/20250904_115056-276cf8b5/summary.json
… artifacts/runs/20250904_124542-8b916e8c/performance.json
… artifacts/runs/20250904_124542-8b916e8c/summary.json
… artifacts/runs/20250904_130556-850d74a8/performance.json
… artifacts/runs/20250904_130556-850d74a8/summary.json